requirements.txt
Django==4.2.27
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
nltk==3.9.2
jieba==0.42.1
regex==2026.1.15
joblib==1.5.3
tqdm==4.67.1
requests==2.32.3
python-dotenv==1.0.1